pyproject.toml
[build-system]
build-backend = "uv_build"
requires = [ "uv-build" ]

[project]
name = "skane-trails-checker"
version = "0.1.0"
description = "A FastAPI + Expo app for tracking hiking trails and foraging spots in Skåne, Sweden"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14.2"
classifiers = [
  "Private :: Do Not Upload",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "fastapi>=0.115.0",
  "geopy>=2.4.0",
  "google-cloud-firestore>=2.19.0",
  "google-cloud-secret-manager>=2.20.0",
  "gpxpy>=1.6.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.0",
  "rdp>=0.8",
  "requests>=2.32.5",
  "uvicorn>=0.34.0",
]
optional-dependencies.dev = [
  "deptry>=0.24.0",
  "google-cloud-secret-manager>=2.20.0",
  "pre-commit>=3.6.0",
  "pyproject-fmt>=2.11.0",
  "ruff>=0.11.0",
  "skane-trails-checker[test]",
  "ty>=0.0.5",
]
optional-dependencies.test = [
  "httpx>=0.28.0",
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.12.0",
  "pytest-xdist>=3.5.0",
]

[dependency-groups]
dev = [
  "skane-trails-checker[dev]",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py313"  # Ruff doesn't support py314 yet, py313 is compatible
line-length = 120
indent-width = 4
cache-dir = ".cache/.ruff_cache"
format.indent-style = "space"
format.quote-style = "double"
format.line-ending = "lf"
format.skip-magic-trailing-comma = true
lint.select = [
  "ALL", # include all the rules, including new ones
]
lint.ignore = [
  "ANN401",  # Any type (acceptable for dynamic types like UploadedFile)
  "BLE001",  # Blind exception (acceptable with logger.warning for user-facing messages)
  "COM812",  # Trailing comma conflicts with formatter
  "D",       # Docstring rules (too verbose for now)
  "DTZ007",  # Datetime without timezone (not critical for this use case)
  "E501",    # Line too long (handled by formatter)
  "INP001",  # Implicit namespace package (dev-tools pattern)
  "PLC0415", # Import not at top-level (needed for optional dependencies like rdp/numpy)
  "S113",    # Request without timeout (not critical for weather API)
  "S311",    # Random for non-crypto (not used for security purposes)
  "T201",    # Print statements (used in dev-tools scripts)
  "TRY300",  # Consider else block (style preference)
]
lint.per-file-ignores."tests/**/*.py" = [
  "ANN001",  # Missing type annotation for function argument (pytest fixtures have dynamic types)
  "ANN201",  # Missing return type annotation (test methods return None)
  "ARG001",  # Unused function argument (fixtures used for side effects)
  "ARG002",  # Unused method argument (fixtures used for side effects)
  "PLR2004", # Allow magic values in tests (self-documenting with comments)
  "S101",    # Allow assert in tests (required for pytest)
]
lint.flake8-quotes.inline-quotes = "double"
lint.isort.combine-as-imports = true
lint.isort.split-on-trailing-comma = false

[tool.deptry]
known_first_party = [ "api", "skane-trails-checker" ]
ignore_notebooks = true
pep621_dev_dependency_groups = [ "dev" ]
# Exclude Cloud Function code (has its own requirements.txt) and venv
exclude = [ "infra/modules/backup/function", ".venv", "venv", "htmlcov" ]
package_module_name_map.skane-trails-checker = "app"
package_module_name_map.fastapi = "fastapi"
# numpy is a transitive dependency used via rdp for coordinate simplification
per_rule_ignores.DEP003 = [ "numpy" ]
# These are local packages within the app/ directory
per_rule_ignores.DEP001 = [ "api", "functions", "resources" ]
# Test dependencies are in optional-dependencies.test group
per_rule_ignores.DEP002 = [ "pytest", "pytest-cov", "pytest-mock", "pytest-xdist", "uvicorn", "httpx" ]

[tool.pyproject-fmt]
keep_full_version = true

[tool.pytest]
ini_options.cache_dir = ".cache/.pytest_cache"
# Test discovery patterns
ini_options.python_files = "test_*.py"
ini_options.python_classes = "Test*"
ini_options.python_functions = "test_*"
# Test paths
ini_options.testpaths = [ "tests" ]
# Output options
ini_options.addopts = [
  "--pythonwarnings=error",
  "-v",
  "--strict-markers",
  "--tb=short",
  "--cov=app",
  "--cov=api",
  "--cov-report=term-missing",
  "--cov-report=html",
]
ini_options.log_cli = true
ini_options.log_cli_level = "INFO"
ini_options.log_cli_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
# Filter specific warnings from external libraries
ini_options.filterwarnings = [
  # Ignore NumPy 2.0 deprecation warning triggered by rdp library
  # The rdp library uses 2D vectors with numpy.cross which is deprecated in NumPy 2.0
  # This needs to be fixed in the rdp library upstream
  "ignore:Arrays of 2-dimensional vectors are deprecated.*:DeprecationWarning",
]

[tool.coverage]
run.branch = true
run.omit = [ "app/functions/bootstrap_trails.py", "dev-tools/*" ]
run.source = [ "api/", "app/" ]
report.exclude_lines = [
  "if __name__ == .__main__.:",
  "pragma: no cover",
]
report.fail_under = 85
report.precision = 2
report.skip_empty = true
